requirement.txt
Django== 4.2
django-cors-headers==4.0.0
djangorestframework
psycopg2